+++ Example/Example/APIClient.swift
@@ -13,7 +13,7 @@ import StripeTerminal
 class APIClient: NSObject, ConnectionTokenProvider {
 
     // This comes from `AppDelegate.backendUrl`, set URL there
-    var baseURLString: String? = nil
+    var baseURLString: String?
 
     private var baseURL: URL {
         if let urlString = self.baseURLString, let url = URL(string: urlString) {
@@ -32,7 +32,7 @@ class APIClient: NSObject, ConnectionTokenProvider {
                 switch responseJSON.result {
                 case .success(let json as [String: AnyObject]) where json["secret"] is String:
                     completion((json["secret"] as! String), nil)
-                case .success(_),
+                case .success,
                      .failure where responseJSON.response?.statusCode == 402:
                     let description = responseJSON.data.flatMap({ String(data: $0, encoding: .utf8) })
                         ?? "Failed to decode connection token"
@@ -40,7 +40,7 @@ class APIClient: NSObject, ConnectionTokenProvider {
                                         code: 1,
                                         userInfo: [
                                             NSLocalizedDescriptionKey: description
-                        ])
+                    ])
                     completion(nil, error)
                 case .failure(let error):
                     completion(nil, error)
@@ -50,20 +50,19 @@ class APIClient: NSObject, ConnectionTokenProvider {
 
     // MARK: Endpoints for App
 
-
     /// Create PaymentIntent using https://github.com/stripe/example-terminal-backend
     ///
     /// - Parameters:
     ///   - params: parameters for PaymentIntent creation
     ///   - completion: called with result: either PaymentIntent client_secret, or the error
-    func createPaymentIntent(_ params: PaymentIntentParameters, completion: @escaping (Swift.Result<String, Error>) -> ()) {
+    func createPaymentIntent(_ params: PaymentIntentParameters, completion: @escaping (Swift.Result<String, Error>) -> Void) {
         let url = self.baseURL.appendingPathComponent("create_payment_intent")
         Alamofire.request(url, method: .post,
                           parameters: [
                             "amount": params.amount,
                             "currency": params.currency,
-                            "description": params.statementDescriptor ?? "Example PaymentIntent",
-            ])
+                            "description": params.statementDescriptor ?? "Example PaymentIntent"
+        ])
             .validate(statusCode: 200..<300)
             .responseJSON { responseJSON in
                 switch responseJSON.result {
@@ -73,15 +72,15 @@ class APIClient: NSObject, ConnectionTokenProvider {
                         return
                     }
                     fallthrough
-                case .success(_),
+                case .success,
                      .failure where responseJSON.response?.statusCode == 402:
                     let description = responseJSON.data.flatMap({ String(data: $0, encoding: .utf8) })
                         ?? "Failed to create PaymentIntent"
                     let error = NSError(domain: "example",
                                         code: 4,
                                         userInfo: [
-                                            NSLocalizedDescriptionKey: description,
-                        ])
+                                            NSLocalizedDescriptionKey: description
+                    ])
                     completion(.failure(error))
                 case .failure(let error):
                     completion(.failure(error))
@@ -104,8 +103,8 @@ class APIClient: NSObject, ConnectionTokenProvider {
                     let error = NSError(domain: "example",
                                         code: 2,
                                         userInfo: [
-                                            NSLocalizedDescriptionKey: description,
-                                            ])
+                                            NSLocalizedDescriptionKey: description
+                    ])
                     completion(error)
                 case .failure(let error):
                     completion(error)
@@ -113,7 +112,7 @@ class APIClient: NSObject, ConnectionTokenProvider {
         }
     }
 
-    func attachPaymentMethod(_ paymentMethodId: String, completion: @escaping ([String: AnyObject]?, Error?) -> ()) {
+    func attachPaymentMethod(_ paymentMethodId: String, completion: @escaping ([String: AnyObject]?, Error?) -> Void) {
         let url = self.baseURL.appendingPathComponent("attach_payment_method_to_customer")
         Alamofire.request(url, method: .post,
                           parameters: ["payment_method_id": paymentMethodId])
@@ -122,15 +121,15 @@ class APIClient: NSObject, ConnectionTokenProvider {
                 switch responseJSON.result {
                 case .success(let json as [String: AnyObject]):
                     completion(json, nil)
-                case .success(_),
+                case .success,
                      .failure where responseJSON.response?.statusCode == 402:
                     let description = responseJSON.data.flatMap({ String(data: $0, encoding: .utf8) })
                         ?? "Failed to decode PaymentMethod & Customer"
                     let error = NSError(domain: "example",
                                         code: 3,
                                         userInfo: [
-                                            NSLocalizedDescriptionKey: description,
-                                            ])
+                                            NSLocalizedDescriptionKey: description
+                    ])
                     completion(nil, error)
                 case .failure(let error):
                     completion(nil, error)