+++ Example/Example/DeviceTypeViewController.swift
@@ -12,7 +12,7 @@ import Static
 
 class DeviceTypeViewController: TableViewController {
 
-    var onSelectedDevice: (DeviceType) -> () = { _ in }
+    var onSelectedDevice: (DeviceType) -> Void = { _ in }
     let deviceType: DeviceType
 
     init(deviceType: DeviceType) {
@@ -27,6 +27,7 @@ class DeviceTypeViewController: TableViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        self.addKeyboardDisplayObservers()
         updateContent()
     }
 
@@ -36,14 +37,14 @@ class DeviceTypeViewController: TableViewController {
                 Row(text: "Chipper 2X", selection: { [unowned self] in
                     self.onSelectedDevice(.chipper2X)
                     self.navigationController?.popViewController(animated: true)
-                }, accessory: (deviceType == .chipper2X ? .checkmark : .none)),
-                ], footer: "Discover the BBPOS Chipper 2X BT.\n\nNote: the Stripe Terminal SDK can discover supported card readers automatically – you don't need to connect to the reader in the iOS Settings > Bluetooth page."),
+                    }, accessory: (deviceType == .chipper2X ? .checkmark : .none))
+            ], footer: "Discover the BBPOS Chipper 2X BT.\n\nNote: the Stripe Terminal SDK can discover supported card readers automatically – you don't need to connect to the reader in the iOS Settings > Bluetooth page."),
             Section(header: "", rows: [
                 Row(text: "Verifone P400", selection: { [unowned self] in
                     self.onSelectedDevice(.verifoneP400)
                     self.navigationController?.popViewController(animated: true)
-                    }, accessory: (deviceType == .verifoneP400 ? .checkmark : .none)),
-                ], footer: "Discover the Verifone P400.\n\nNote: Readers must be registered to your account via the Stripe API or Dashboard before they're discoverable."),
+                    }, accessory: (deviceType == .verifoneP400 ? .checkmark : .none))
+            ], footer: "Discover the Verifone P400.\n\nNote: Readers must be registered to your account via the Stripe API or Dashboard before they're discoverable.")
         ]
     }