+++ Example/StripeTerminal.framework/Headers/SCPCharge.h
@@ -84,7 +84,7 @@ typedef NS_ENUM(NSUInteger, SCPChargeStatus) {
 /**
  You cannot directly instantiate this class.
  */
-- (instancetype)new NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
 
 @end