+++ Example/StripeTerminal.framework/Headers/SCPReaderSoftwareUpdate.h
@@ -72,7 +72,7 @@ NS_SWIFT_NAME(ReaderSoftwareUpdate)
 /**
  You cannot directly instantiate this class.
  */
-- (instancetype)new NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
 
 @end